+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Renamed task file from `Merge-CodeCoverageFiles.pester.build.ps1` to `JaCoCo.coverage.build.ps1`.
 - Move task _Convert_Pester_Coverage_ to task file `JaCoCo.coverage.build.ps1`.
+- Resolve-Dependency.ps1: fix MinimumPSDependVersion comparison.
 
 ## [0.110.1] - 2021-04-08